py-cassandra/requirements.txt
cassandra-driver==3.19.0
numpy==1.17.2
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
